+++ types/User/User.d.ts
@@ -1,0 +1,57 @@
+import { IObjectWithId } from '../base';
+
+/**
+ * The User object can be used to manage your own account and to create child accounts.
+ * Only a Production mode API key can be used to make requests against the Users API.
+ *
+ * Balance and recharge values on User objects are expressed in higher precision USD.
+ *
+ * @see https://www.easypost.com/docs/api/node#user-object
+ */
+export declare interface IUser extends IObjectWithId<'User'> {
+  /**
+   * The ID of the parent user object.
+   * Top-level users are defined as users with no parent
+   */
+  parent_id: string;
+
+  /**
+   * First and last name required
+   */
+  name: string;
+
+  /**
+   * Required
+   */
+  email: string;
+
+  /**
+   * Optional
+   */
+  phone_number: string;
+
+  /**
+   * Formatted as string "XX.XXXXX"
+   */
+  balance: string;
+
+  /**
+   * USD formatted dollars and cents, delimited by a decimal point
+   */
+  recharge_amount: string;
+
+  /**
+   * USD formatted dollars and cents, delimited by a decimal point
+   */
+  secondary_recharge_amount: string;
+
+  /**
+   * Number of cents USD that when your balance drops below, we automatically recharge your account with your primary payment method.
+   */
+  recharge_threshold: string;
+
+  /**
+   * All associated children
+   */
+  children: IUser[];
+}
